Fix pre-commit

+++ chia/wallet/wallet_state_manager.py
@@ -737,7 +737,7 @@ class WalletStateManager:
                 self, self.main_wallet, name="NFT Wallet", in_transaction=True
             )
             self.log.error("Create NFT wallet")
-            wallet_id = nft_wallet.wallet_id
+            wallet_id = uint32(nft_wallet.wallet_id)
             wallet_type = WalletType.NFT
 
         return wallet_id, wallet_type
